refactor(all): split css and architecture
