+++ src/components/App/App.sass
@@ -2,7 +2,7 @@ html
   height: 100%
 
 body
-  padding-top: 70px
+  padding-top: 50px
   min-height: 100%
   background-color: #2d3742
   color: #fff
