+++ src/components/header/Header.sass
@@ -12,6 +12,13 @@
     font-size: 1.6em
     padding: 12px 15px
     vertical-align: middle
+    a
+      &:hover
+        color: #528CC1
+      &:focus
+        color: #fff
+        text-decoration: none
+        outline: none
 
     .icon-site
       color: #528CC1
