+++ src/components/paginator/Paginator.sass
@@ -1,0 +1,12 @@
+.pager
+  margin: 0
+  padding: 0
+
+  li  a
+    background: #222931
+    color: #fff
+    border: 1px solid rgba(255,255,255, 0.1)
+    &:hover
+      background-color: #528CC1
+      color: #fff
+      cursor: pointer
