+++ src/components/repository/Repository.sass
@@ -1,89 +1,2 @@
 .repository
   height: 100%
-
-  .leftpanel
-    padding: 5px
-    .authorInformations
-      text-align: center
-      padding-top: 5px
-      img
-        width: 80%
-        height: 80%
-
-    .basicInformations
-      margin-top: 10px
-
-  .mainpanel
-    padding: 5px
-    .contributors
-      h2
-        margin-top: 0
-      .contributor
-        text-align: center
-        img
-          width: 75px
-          height: 75px
-      .pager > li > a
-        background: #222931
-        color: #fff
-        border: 1px solid rgba(255,255,255, 0.1)
-        &:hover
-          background-color: #528CC1
-          color: #fff
-          cursor: pointer
-
-  .header
-    margin-bottom: 10px
-    h1
-      line-height: 35px
-      text-align: center
-      a
-        color: white
-        &:hover
-          text-decoration: none
-          color: white
-
-  .metrics
-    .menu-panel
-      display: table
-      table-layout: fixed
-      padding: 0
-      width: 100%
-      list-style: none
-      .button
-        position: relative
-        display: table-cell
-        text-transform: uppercase
-        font-size: medium
-        line-height: 40px
-        box-shadow: 2px 0 5px 0px rgba(0, 0, 0, 0.5)
-        text-align: center
-        background-color: #222931
-        border-right-style: solid
-        border-bottom-style: solid
-        border-color: rgba(255,255,255, 0.1)
-        border-width: 1px
-        &:hover
-          background-color: #528CC1
-          color: #fff
-          cursor: pointer
-        &::after
-          content: " "
-          position: absolute
-          display: block
-          left: 0
-          bottom: 0
-          width: 100%
-          height: .25em
-          background-color: #528CC1
-          opacity: 0
-
-      .button-selected
-        color: #528CC1
-        font-weight: bold
-        &::after
-          opacity: 1
-
-    .highcharts-container
-      margin: 10px 0 0 0
-      box-shadow:  2px 0 5px 0px rgba(0, 0, 0, 0.5)
