+++ src/components/repository/containers/ContributorsContainer.jsx
@@ -1,8 +1,8 @@
 import React, { Component } from 'react'
 import LoadingWrapper from './../../loader/LoadingWrapper'
 
-import Contributors from './../presentational/mainpanel/Contributors'
-import Paginator from './../presentational/mainpanel/Paginator'
+import Contributors from './../presentational/mainpanel/contributors/Contributors'
+import Paginator from './../../paginator/Paginator'
 import GithubApiService from '../../../services/GithubApiService'
 import RepositoryConstants from './../../../constants/RepositoryConstants'
 import CacheService from './../../../services/CacheService'
@@ -50,18 +50,16 @@ export default class ContributorsContainer extends Component {
     })
   }
 
-  _handlePaginationClick = (next) => {
-    return () => {
-      this.setState({loaded: false})
-      this.fetchContributors(next.page, next.per_page)
-    }
+  _handlePaginationClick = (pagination) => {
+    this.setState({loaded: false})
+    this.fetchContributors(pagination.page, pagination.per_page)
   }
 
   render() {
     return (
       <LoadingWrapper loaded={this.state.loaded} >
         <Contributors contributors={this.state.contributors}>
-          <Paginator pagination={this.state.pagination} onPaginationClick={this._handlePaginationClick} />
+          <Paginator pagination={this.state.pagination} onClick={this._handlePaginationClick} />
         </Contributors>
       </LoadingWrapper>
     )
