+++ src/components/repository/presentational/Header.sass
@@ -1,0 +1,10 @@
+.header
+  margin-bottom: 5px
+  h1
+    line-height: 30px
+    text-align: center
+    a
+      color: white
+      &:hover
+        text-decoration: none
+        color: white
