+++ src/components/repository/presentational/LeftPanel.jsx
@@ -1,5 +1,7 @@
 import React from 'react'
 
+import './LeftPanel.sass'
+
 const LeftPanel = ({children}) => {
   return (
     <div className="col-md-3 leftpanel">
