+++ src/components/repository/presentational/LeftPanel.sass
@@ -1,0 +1,2 @@
+.leftpanel
+  padding: 5px
