+++ src/components/repository/presentational/MainPanel.jsx
@@ -1,5 +1,7 @@
 import React from 'react'
 
+import './MainPanel.sass'
+
 const MainPanel = ({children}) => {
   return (
     <div className="col-md-9 mainpanel">
