+++ src/components/repository/presentational/MainPanel.sass
@@ -1,0 +1,13 @@
+.mainpanel
+  padding: 5px
+
+  .contributors
+    padding: 0 20px 0 20px
+    margin: auto
+    h2
+      margin-top: 0
+
+    .contributor
+      img
+        width: 75px
+        height: 75px
