+++ src/components/repository/presentational/leftpanel/AuthorInformations.sass
@@ -1,0 +1,6 @@
+.authorInformations
+  text-align: center
+  padding-top: 5px
+  img
+    width: 80%
+    height: 80%
