+++ src/components/repository/presentational/leftpanel/SimpleInformations.sass
@@ -1,0 +1,2 @@
+.simpleInformations
+  margin-top: 10px
