+++ src/components/repository/presentational/mainpanel/metrics/TimelineCommits.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import ReactHighcharts from 'react-highcharts/bundle/ReactHighcharts'
+import ReactHighcharts from '../../../../../../node_modules/react-highcharts/bundle/ReactHighcharts'
 
 const TimelineCommits = ({data}) => {
   const config = {
