+++ src/components/repository/presentational/mainpanel/metrics/conf/chartsConf.js
