+++ src/components/searchbar/SearchBar.sass
@@ -1,5 +1,4 @@
 .searchpanel
-
   .input-group-addon
     border-radius: 0
 
@@ -12,7 +11,6 @@
     margin: 0
     padding: 0
     text-shadow: 0 1px 0.1em rgba(0,0,0,.4)
-    z-index: 1
 
     .li
       width: 100%
@@ -23,13 +21,12 @@
       text-decoration: none
       &:hover
         @extend .li-hover
-      .pull-right
-        padding: 0
-        margin: 0
+
       .label
         margin: 0 0 0 3px
         .glyphicon-star
           color: gold
+
     .nbResults
       width: 100%
       background-color: ghostwhite
@@ -37,11 +34,8 @@
       color: black
       text-align: center
 
-
     .li-hover
       background-color: #528CC1
       cursor: pointer
       a
         color: white
-
-
